ANN_BENCH: Fix segfault in CAGRA wrapper when moving its graph (#733)
When the graph in CAGRA wrapper needs to be relocated to a different memory resource, the old one is destructed before the data is copied. This leads to a segfault during copy. The fix is to retain the old graph pointer until the data is copied.

This error is likely to happen when one runs CAGRA search benchmark sweeping over multiple values of `graph_memory_type`.

Authors:
  - Artem M. Chirkin (https://github.com/achirkin)

Approvers:
  - Tamas Bela Feher (https://github.com/tfeher)

URL: https://github.com/rapidsai/cuvs/pull/733