Update score calculation for CAGRA-Q instance selection (#938)
The current CAGRA-Q instance selection criterion is the same as the one for the standard CAGRA, which is not always optimal for CAGRA-Q. This PR updates the criterion for CAGRA-Q to improve the throughput when `team_size=AUTO`.


The size (Byte) of each vector is smaller in a dataset compressed with CAGRA-Q compared to an uncompressed one. Because of this, we may be able to improve throughput by using a smaller `team_size`. This PR updates the scoring method for selecting a CAGRA-Q instance to take that into account. Based on my performance tests for SIFT, GloVe, GIST, NYTimes, and OpenAI 5M, the updated scoring method avoided selecting the worst `team_size` values, unlike the current method.

Authors:
  - tsuki (https://github.com/enp1s0)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/cuvs/pull/938