Reuse `<version_config.h>` instead of separate version constants in `c_api.h` (#1003)
This commit started off trying to fix an error in how version strings are replaced in the C API header. (Bumping the version to 25.08 causes the minor version to be written as "08" instead of "8", producing an invalid octal number.)

The octal problem was already solved as part of RAPIDS CMake, which is used in cuVS to generate a `version_configi.h` with the corrected version values.  This commit removes the old version constants, and switches to using what's provided by infrastructure.

Authors:
  - MithunR (https://github.com/mythrocks)
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)

URL: https://github.com/rapidsai/cuvs/pull/1003