[Review][Java] Fix random segabort/segfault/double free problems (#1045)
The cuVS C API we are leveraging uses a common C pattern for "object creation": you pass down a pointer, and that gets populated. Often, this is a **, and memory gets allocated on the other side:
```
create(voi**ptrPtr) {
   *ptrPtr = new Something;
}
void** ptrPtr = // allocate space for a pointer
create(ptrPtr);
void* ptr = *ptrPtr;
```
In Panama, this means 
- allocating a small pointer size MemorySegment, very short lived
- call `create`
- extract a MemorySegment from the MemorySegment

Existing code sometimes skipped the extraction bit (the dereference), and/or created the structure directly instead of using the C `create` functions; this lead to some segfaults (addressing the wrong memory/wrong size) and to some double frees (the JVM freed MemorySegments we allocated via an arena, but we also called the `destroy` functions, so the C side tried to free them too).

Plus, this PR tries to reduce the lifetime of some allocations, avoiding to use the `CuVSResources` arena when it is clearly not needed -- but I did not want to expand the scope too much, so I just did the more obvious ones. Optimizing memory allocation is still a WIP, especially on the search side, but this I think will deserve its own PR.

These changes makes the test suite pass reliably - I have not seen a segmentation fault over hundreds of executions locally, and lowers the native memory occupied a little.

Authors:
  - Lorenzo Dematté (https://github.com/ldematte)
  - Ben Frederickson (https://github.com/benfred)

Approvers:
  - Chris Hegarty (https://github.com/ChrisHegarty)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cuvs/pull/1045