hnsw::from_cagra: avoid allocating the graph twice (#1057)
When CAGRA graph is not host accessible, HNSW::from_cagra routine copies the data from it to the base layer of HNSW in two stages: first to a host matrix, then to the HNSW index object. As a result, the graph, which can be huge, is allocated in the host memory twice.
For big enough graphs this manifests itself in the program being killed by the linux kernel at the end of the index building process due to not enough system memory.

This PR avoids the staging graph buffer by using cudaMemcpy2DAsync to copy the data in an appropriate format in one go.

Authors:
  - Artem M. Chirkin (https://github.com/achirkin)
  - Ben Frederickson (https://github.com/benfred)

Approvers:
  - Micka (https://github.com/lowener)

URL: https://github.com/rapidsai/cuvs/pull/1057