Fix go CI failures (#1114)
Fixes #1113.

We are currently seeing CI failures on the go-build: with an error message like `build constraints exclude all Go files in /__w/cuvs/cuvs/go`

Since we have `import "C"`  declarations  we need to make sure that we have CGO_ENABLED set. The difference in the failing vs non-failing CI runs seem to be that we are installing the `nocgo` version of `_go_select` in the failing CI runs, instead of the `cgo` version in the successful runs. 

Fix by installing `_go_select *=cgo` in the conda environments

Authors:
  - Ben Frederickson (https://github.com/benfred)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cuvs/pull/1114