IVF-PQ coarse search: fix integer overflow and avoid excessive batch sizes (#999)
When the IVF-PQ batch size and the number of cluster are big enough, the GEMM output matrix size may exceed the `uint32_t` limits. This manifests as an illegal memory access popping up in further stages of IVF-PQ (due to everything being as async as possible).

This PR fixes the bug by converting the input arguments to 64-bits before multiplying them.

On top of the overflow fix, this PR also limits the batch size when it may hurt performance: (1) within IVF-PQ to avoid out-of-memory error (2) in CAGRA graph build to avoid switching to the large memory resource (which is typically backed by managed memory).

Authors:
  - Artem M. Chirkin (https://github.com/achirkin)
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - Tamas Bela Feher (https://github.com/tfeher)

URL: https://github.com/rapidsai/cuvs/pull/999