Expose graph and dataset accessors for CAGRA to C/Python (#1086)
Add the ability to get the graph and dataset from a CAGRA index to the c-api and python apis, as well as being able to reconstruct the cagra index from a graph and dataset.

The eventual goal here is to be more flexible in terms of allowing other serialization formats. Rather than supporting every format inside of cuvs, by exposing the raw data needed to recreate a cagra index - we can let consumers of cuvs decide how they want to serialize an index.

Authors:
  - Ben Frederickson (https://github.com/benfred)
  - MithunR (https://github.com/mythrocks)

Approvers:
  - MithunR (https://github.com/mythrocks)
  - Corey J. Nolet (https://github.com/cjnolet)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cuvs/pull/1086