Add error checking for our C example code (#1171)
We weren't consistently checking the return value of cuda and cuvs functions in our C examples.

This caused problems when we tried to adapt the SQ code to an example, and it was returning invalid results. The root cause was because a function was failing, but we didn't notice because the pattern has been to ignore the error codes.

Enforce best practices and check the result of every cuvs and cuda function call in our c examples.

Also:
* enable -Wall and -Werror to the c examples for extra compile time checks
* add examples to CODEOWNERS
*  ensure that clang-format runs on the example files

Authors:
  - Ben Frederickson (https://github.com/benfred)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)
  - Gil Forsyth (https://github.com/gforsyth)

URL: https://github.com/rapidsai/cuvs/pull/1171