(Docs) Switch to sphinx-autoapi with stub-based hipvs API refs (#54)
* (Docs) Switch to sphinx-autoapi with stub-based hipvs API refs

Replaced autodoc with sphinx-autoapi to generate Python API docs from
pybind11-stubgen stubs. Added a README with stub generation steps and new
.rst files for common, random, and sparse modules. Updated Sphinx
requirements to include sphinx-autoapi and its dependencies.

The main motivation for this change is that sphinx-autodoc requires
successfully importing `hipvs` in the build environment (e.g.,
ReadTheDocs). Since that environment lacks the full ROCm stack and is
CPU-only, importing fails. Using autoapi with checked-in stubs avoids
this dependency and ensures docs can be built consistently.

* Add documnetation for cuvs.common module

* Bake the rust documentation HTML files into the doc sources

This approach ensures that we don't need to depend on cargo when building
docs.