[Java][C] Expose GPUInfo (#1267)
cuvs-java already contains a public `GPUInfo` class, but methods to retrieve the information and fill it are internal. 
This PR exposes them through and interface, `GPUInfoProvider`. It also separates immutable data related to a GPU (which is kept in `GPUInfo`) from transient resources-related data and counters (at the moment, only the amount of free memory, which is kept in the new `CuVSResourcesInfo`).

The change let you query transient data at a later moment; to do this, we need to find the device ID associated with a `CuVSResource` object. The change to the C API exposes the raft function that does it.

Authors:
  - Lorenzo Dematté (https://github.com/ldematte)
  - MithunR (https://github.com/mythrocks)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cuvs/pull/1267