Don't use check_cuvs in __dealloc__
This was causing warnings such as:

TypeError: 'NoneType' object is not callable
Exception ignored in: 'cuvs.common.resources.Resources.__dealloc__'

during finalization of a program due to check_cuvs being destroyed
before the call to dealloc.

Secondly, check_cuvs raises an exception on error which will be ignored
as stated in the warning. Instead, just log the error.