Add cagra.extend bindings for python (#1324)
Add the cagra.extend functionality for python.

Note that this PR also removes the `return_dataset` parameter from the C-API in favour of the cuvsCagraIndexGetDataset function.  This makes it optional to get the extended dataset in c/python, and also doesn't perform a copy.

Authors:
  - Ben Frederickson (https://github.com/benfred)
  - Corey J. Nolet (https://github.com/cjnolet)

Approvers:
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/cuvs/pull/1324