Fix float to integral type cast OOR UB
`cuvs::spatial::knn::detail::utils::mapping` converts values between types
and applies scaling when converting to/from float and integral types.
It is undefined behaivor, when casting from float to integral values,
the float value is out-of-range of the integral type. This change avoids
the UB by clamping the value to the valid range before casting.

Fixes: AIOSS-2153