Update the documentation with latest instructions (#105)
* Update the documentation with latest instructions

When building both the python and rust bindings from source, it is
required that we have already installed the hipRaft C++ library and
related CMake module related files into the same environment from
where CMake can pick it up when building the bindings.

This PR just adds these instructions and details to the docs.

* Address PR comment