[AMD][Tests] Enable Neighbosr tests
While running the tests, there was often a seg fault when
attempting to allocate over 2 petabytes of memory on a GPU.
This is because there is a test where num_ranks == 0,
which causes a division by zero when calculating the sizes
needed for each GPU in a batched build.

There was another instance of this when the number of clusters was
less than the number of ranks. This caused another division by zero
and a test failure.

Replacing the deprecated `omp_set_nested` with `omp_set_max_active_levels`
reduced the number of test failures until the actual error was discovered,
so we are keeping it for now.